[FBS-129] - refactor: hook created to remove duplication
