+++ src/pages/settings/components/dashboard-panel/index.tsx
@@ -1,60 +1,25 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { VariantType, useSnackbar } from "notistack";
 import sign from "jwt-encode";
-import jwtDecode from "jwt-decode";
 
 import { IPayloadDashboardForm } from "pages/settings/types/IPayloadDashboardForm";
 
-import { createToken } from "services/resources/token";
 import { setDashboardUrl } from "services/api";
 
 import { LocalStorage } from "shared/utils/base-storage/local-storage";
-import { SessionStorage } from "shared/utils/base-storage/session-storage";
 
 import * as S from "./styles";
+import { useForm } from "pages/settings/hooks/useForm";
 
 export const DashboardPanel: React.FC = () => {
   const navigate = useNavigate();
-  const { enqueueSnackbar } = useSnackbar();
+  const { onSetToken, showNotification } = useForm();
 
   const [isLoadingDashboard, setIsLoadingDashboard] = useState(false);
   const urlMetabase = process.env.REACT_APP_METABASE_SITE_URL as string;
   const secretKeyMetabase = process.env.REACT_APP_METABASE_SECRET_KEY as string;
   const dashboardMetabaseNumber = process.env
     .REACT_APP_METABASE_DASHBOARD_NUMBER as string;
-
-  function showNotification(message: string, variant: VariantType) {
-    enqueueSnackbar(message, {
-      autoHideDuration: 4000,
-      variant,
-    });
-  }
-
-  function getUserId() {
-    const token = SessionStorage.getInstance().getValueByKey<string>("TOKEN");
-
-    if (!token) {
-      return "";
-    }
-
-    const decoded = jwtDecode(token);
-    SessionStorage.getInstance().setValue("TOKEN", token);
-    return decoded;
-  }
-
-  async function onHandleToken() {
-    const userId = getUserId() as string;
-    const token = await createToken(userId);
-
-    if (!token) {
-      const message = "Erro no retorno do Token. Por favor tentar novamente!";
-      showNotification(message, "error");
-      return;
-    }
-
-    SessionStorage.getInstance().setValue("TOKEN", token);
-  }
 
   async function onSubmitDashboard(payload: IPayloadDashboardForm) {
     try {
@@ -75,7 +40,7 @@ export const DashboardPanel: React.FC = () => {
         `${payload.metabaseSiteUrl}/embed/dashboard/${token}#theme=night&bordered=true&titled=true`
       );
 
-      await onHandleToken();
+      await onSetToken();
 
       const message = "Sucesso ao conectar com o servidor";
       showNotification(message, "success");
