+++ src/pages/settings/components/flowbuild-panel/index.tsx
@@ -1,19 +1,15 @@
-import { VariantType, useSnackbar } from "notistack";
-import jwtDecode from "jwt-decode";
-
 import { IPayloadForm } from "pages/settings/types/IPayloadForm";
 
 import { LocalStorage } from "shared/utils/base-storage/local-storage";
-import { SessionStorage } from "shared/utils/base-storage/session-storage";
 
-import { createToken } from "services/resources/token";
 import { healthcheck } from "services/resources/settings";
 import { setBaseUrl } from "services/api";
 
 import * as S from "./styles";
+import { useForm } from "pages/settings/hooks/useForm";
 
 export const FlowbuildPanel: React.FC = () => {
-  const { enqueueSnackbar } = useSnackbar();
+  const { onSetToken, showNotification } = useForm();
 
   const defaultValue = {
     url:
@@ -35,38 +31,6 @@ export const FlowbuildPanel: React.FC = () => {
     },
   };
 
-  function showNotification(message: string, variant: VariantType) {
-    enqueueSnackbar(message, {
-      autoHideDuration: 4000,
-      variant,
-    });
-  }
-
-  function getUserId() {
-    const token = SessionStorage.getInstance().getValueByKey<string>("TOKEN");
-
-    if (!token) {
-      return "";
-    }
-
-    const decoded = jwtDecode(token);
-    SessionStorage.getInstance().setValue("TOKEN", token);
-    return decoded;
-  }
-
-  async function onHandleToken() {
-    const userId = getUserId() as string;
-    const token = await createToken(userId);
-
-    if (!token) {
-      const message = "Erro no retorno do Token. Por favor tentar novamente!";
-      showNotification(message, "error");
-      return;
-    }
-
-    SessionStorage.getInstance().setValue("TOKEN", token);
-  }
-
   async function onSubmitServer(payload: IPayloadForm) {
     let url = payload.url;
 
@@ -76,7 +40,7 @@ export const FlowbuildPanel: React.FC = () => {
       LocalStorage.getInstance().setValue("SERVER_URL", url);
       LocalStorage.getInstance().setValue("SERVER_PORT", payload.port);
       setBaseUrl(url);
-      await onHandleToken();
+      await onSetToken();
 
       const message = "Sucesso ao salvar o servidor";
       showNotification(message, "success");
